Updates to cis129_lab03_coffeeShop.ipynb
Updated the program with two more menu items so the Coffee Shop has 4 items that users can purchase (matcha and croissants). Update the receipt to ensure that it displays the correct number of items ordered and correct values. Changed the wording on the receipt ("purchased" instead of "bought"). Added a thank you message at the end.